+++ package.json
@@ -26,8 +26,7 @@
     "rollup-plugin-node-resolve": "^5.2.0",
     "rollup-plugin-serve": "^1.0.1",
     "rollup-plugin-svelte": "^5.2.3",
-    "rollup-plugin-terser": "^6.1.0",
-    "svelte": "^3.0.0"
+    "rollup-plugin-terser": "^6.1.0"
   },
   "dependencies": {
     "clipboard": "^2.0.6",
@@ -39,6 +38,7 @@
     "dragula": "^3.7.2",
     "moment": "^2.27.0",
     "sirv-cli": "^0.4.4",
+    "svelte": "^3.24.0",
     "svelte-feather-icons": "^3.2.2",
     "svelte-i18n": "^3.0.4",
     "svelte-textarea-autoresize": "^1.0.0",
